chart for city done

+++ assets/js/db.js
@@ -1,14 +1,19 @@
 // LZString from https://github.com/pieroxy/lz-string
 // imported in html
 
-async function writeDataToDb(key, data) {
-    const compressed = await LZString.compressToUTF16(JSON.stringify(data));
+function writeDataToDb(key, data, force = false) {
+    if (!force && key in localStorage) return null;
+    const compressed = LZString.compressToUTF16(JSON.stringify(data));
     localStorage.setItem(key, compressed);
 }
 
-async function getDataFromDb(key) {
-    const data = await LZString.decompressFromUTF16(localStorage.getItem(key));
-    return data;
+function getDataFromDb(key) {
+    const data = LZString.decompressFromUTF16(localStorage.getItem(key));
+    return data ? JSON.parse(data) : null;
+}
+
+function clearDB() {
+    localStorage.clear();
 }
 
 function getLocalStorageSize() {
@@ -28,4 +33,4 @@ function getLocalStorageSize() {
     console.log("Total = " + (_lsTotal / 1024).toFixed(2) + " KB");
 }
 
-export default { writeDataToDb, getDataFromDb, getLocalStorageSize };
+export default { writeDataToDb, getDataFromDb, getLocalStorageSize, clearDB };
